+++ src/blueprints/Do.Blueprints.Service.Application/Core/CoreExtensions.cs
@@ -5,6 +5,7 @@ using Do.Testing;
 using Newtonsoft.Json;
 using Shouldly;
 using System.Diagnostics.CodeAnalysis;
+using System.Reflection;
 using System.Text.RegularExpressions;
 
 namespace Do;
@@ -91,4 +92,44 @@ public static class CoreExtensions
     public static string? ToJsonString(this object? payload) => payload is null ? null : JsonConvert.SerializeObject(payload);
     [return: NotNullIfNotNull("payload")]
     public static object? ToJsonObject(this object? payload) => JsonConvert.DeserializeObject(payload.ToJsonString() ?? string.Empty);
+
+    public static PropertyInfo? PropertyOf<T>(this Stubber _, string name) =>
+        typeof(T).GetProperty(name, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance);
+
+    public static void ShouldBe<T>(this Type source) => source.ShouldBe(typeof(T));
+
+    public static void ShouldBeAbstract(this PropertyInfo source)
+    {
+        var getMethod = source.GetGetMethod(true);
+
+        getMethod.ShouldNotBeNull();
+        getMethod.ShouldBeAbstract();
+    }
+
+    public static void ShouldBeVirtual(this PropertyInfo source)
+    {
+        var getMethod = source.GetGetMethod(true);
+
+        getMethod.ShouldNotBeNull();
+        getMethod.ShouldBeVirtual();
+    }
+
+    public static MethodInfo? MethodOf<T>(this Stubber _, string name) =>
+        typeof(T).GetMethod(name, BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance);
+
+    public static void ShouldBeAbstract(this MethodInfo source)
+    {
+        source.IsAbstract.ShouldBeTrue();
+    }
+
+    public static void ShouldBeVirtual(this MethodInfo source)
+    {
+        source.IsVirtual.ShouldBeTrue();
+    }
+
+    public static void ShouldHaveOneParameter<T>(this MethodInfo source)
+    {
+        source.GetParameters().Length.ShouldBe(1);
+        source.GetParameters().First().ParameterType.ShouldBe<T>();
+    }
 }
